Move openBlockSeriesChunkRefsSetsIterator
